pyproject.toml
[project]
name = "py-for-finance-cookbk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cpi>=2.0.7",
    "google-search-results>=2.4.2",
    "ipykernel>=6.29.5",
    "nbformat>=5.10.4",
    "pandas>=2.2.3",
    "plotly>=6.1.2",
    "polars>=1.30.0",
    "python-dotenv>=1.1.0",
    "yfinance>=0.2.61",
]
